Rename .users package to .user
* 패키지 이름에 복수형을 쓴 것이 부자연스럽다고 생각하여 변경하였음.
