+++ src/main/java/kr/codesquad/cafe/user/User.java
@@ -1,4 +1,4 @@
-package kr.codesquad.cafe.users;
+package kr.codesquad.cafe.user;
 
 import org.springframework.util.Assert;
