+++ src/main/java/kr/codesquad/cafe/user/UserController.java
@@ -1,4 +1,4 @@
-package kr.codesquad.cafe.users;
+package kr.codesquad.cafe.user;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
