+++ src/main/java/kr/codesquad/cafe/user/UserCreationForm.java
@@ -1,4 +1,4 @@
-package kr.codesquad.cafe.users;
+package kr.codesquad.cafe.user;
 
 public class UserCreationForm {
