+++ src/main/java/kr/codesquad/cafe/user/UserRepository.java
@@ -1,4 +1,4 @@
-package kr.codesquad.cafe.users;
+package kr.codesquad.cafe.user;
 
 import org.springframework.stereotype.Repository;
